Refactored
All node-database related stuff now in their separate folder. Updated all files using the old located connection-file.
