+++ src/lib/httpStatusCodes.ts
@@ -11,3 +11,18 @@ export const STATUS_METHOD_NOT_ALLOWED = 405;
 
 /* 5xx Server Error */
 export const STATUS_SERVER_ERROR = 500;
+
+/* Single object with all statuses makes the importing a lot less messy */
+export const HTTP_STATUS = {
+    /* 2xx Success */
+    OK: 200,
+    CREATED: 201,
+
+    /* 4xx Client Error */
+    BAD_REQUEST: 400,
+    FORBIDDEN: 403,
+    NOT_ALLOWED: 405,
+
+    /* 5xx Server Error */
+    SERVER_ERROR: 500
+}
