+++ src/pages/api/v1/bylocation.js
@@ -1,7 +1,7 @@
 import mysql from "mysql2/promise"
 import {z, ZodError} from "zod";
 
-import { getConnectionPool } from "src/lib/database";
+import { getConnectionPool } from "src/lib/database/connection";
 
 import {
     STATUS_OK,
