+++ src/pages/api/v1/index.js
@@ -2,7 +2,7 @@
 import mysql from 'mysql2/promise';
 import { ZodError } from "zod";
 
-import { getConnectionPool } from "src/lib/database";
+import { getConnectionPool } from "src/lib/database/connection";
 import { parseUnit as parseTempUnit } from "src/lib/units/temperature";
 import { parseUnit as parseCondUnit } from "src/lib/units/conductivity";
 import {
